Renamed soundcard file
